Diagramm

+++ war/js/App.js
@@ -1,9 +1,13 @@
 /* global document: false */
 
 function App() {
+	this.locked = false;
     this.assets = 1000.0;
     this.winnings = 0;
     this.DEFAULT_TIME = 0;
+    this.currentIdx = 0;
+    this.paid = 0;
+    this.graphValues = [ [ 'No.', 'Bet', 'Profit'] ]
 }
 
 App.prototype.start = function() {
@@ -98,37 +102,47 @@ App.prototype.printProfit = function(profit) {
 	
 App.prototype.displayProfit = function(profit){
     var div = document.getElementById("profits");
-    div.innerText = "You won "+profit.toFixed(2)+" €";
+    div.innerText = "You won " + profit.toFixed(2) + " â‚¬";
+
+    this.currentIdx++;
+    this.graphValues.push( [  'No ' + this.currentIdx, 
+                              parseFloat(this.winnings.toFixed(2)), 
+                              parseFloat(profit.toFixed(2))
+                            ]);
 
     this.assets = this.assets + profit;
     this.winnings = profit - this.winnings;
-
+    
     div = document.getElementById("winnings");
-    div.innerText = "Profits: "+this.winnings.toFixed(2)+" €";
+    div.innerText = "Profits: "+this.winnings.toFixed(2)+" â‚¬";
 
     this.computeAssets(0);
-
+    this.drawGrapth();
 };
 	
 App.prototype.initiateSequence = function(){
-    document.getElementById("profits").innerHTML="";
-    var table = document.getElementById("eventTable");
-    for(var i = table.rows.length - 1; i > 1; i--)
-    {
-        table.deleteRow(i);
-    }
-
-    document.getElementById("winnings").innerText = "";
-    document.getElementById("paid_money").innerText = "";
-    document.getElementById("profits").innerText = "";
-
-    var waitTime = this.DEFAULT_TIME;
-    var waitTimeInput = document.getElementById("waitingTime");
-    if(waitTimeInput) {
-    	waitTime = Number.parseInt(waitTimeInput.value);
-    }
-    
-    launchWorkflow(waitTime);
+	if (!this.locked) {
+		this.locked = true;
+		
+	    document.getElementById("profits").innerHTML="";
+	    var table = document.getElementById("eventTable");
+	    for(var i = table.rows.length - 1; i > 1; i--)
+	    {
+	        table.deleteRow(i);
+	    }
+	
+	    document.getElementById("winnings").innerText = "";
+	    document.getElementById("paid_money").innerText = "";
+	    document.getElementById("profits").innerText = "";
+	
+	    var waitTime = this.DEFAULT_TIME;
+	    var waitTimeInput = document.getElementById("waitingTime");
+	    if(waitTimeInput) {
+	    	waitTime = Number.parseInt(waitTimeInput.value);
+	    }
+	    
+	    launchWorkflow(waitTime);
+	}
 };
 	
 App.prototype.computeAssets = function(paid){
@@ -136,12 +150,28 @@ App.prototype.computeAssets = function(paid){
     console.log(this.assets, paid);
     var span = document.getElementById("assets");
     span.innerHTML="";
-    var txt = document.createTextNode("Available : "+this.assets.toFixed(2)+" €");
+    var txt = document.createTextNode("Available : "+this.assets.toFixed(2)+" â‚¬");
     span.appendChild(txt);
 
     if (paid > 0) {
         this.winnings = paid;
         var div = document.getElementById("paid_money");
-        div.innerText = "You paid " + paid.toFixed(2) + " €";
+        div.innerText = "You paid " + paid.toFixed(2) + " â‚¬";
     }
 };
+
+App.prototype.drawGrapth = function() {
+    var data = google.visualization.arrayToDataTable(this.graphValues);
+
+    var options = {
+      titlePosition: 'none',
+      legend: { position: 'bottom' },
+      height: 250
+    };
+
+    var chart = new google.visualization.LineChart(document.getElementById('chart_div'));
+
+    chart.draw(data, options);
+    
+    this.locked = false;
+}
